requirements.txt
aiohttp==1.0.5
async-timeout==1.4.0
certifi==2017.7.27.1
chardet==3.0.4
discord.py==0.16.12
idna==2.6
multidict==3.2.0
pycparser==2.18
PyNaCl==1.0.1
requests==2.20.0
six==1.11.0
spotipy==2.4.4
urllib3==1.24.2
websockets==3.4
youtube-dl==2017.8.18
